pytorch/mnist/requirements.txt
# main.py
torch
torchvision

# app.py
Pillow